requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.75.0
flask==3.1.0
flask-cors==5.0.1
werkzeug==3.1.3
jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0 